make repo up-to-date (#6)
* add Issue Template Bug (#228)

Template for adding bugs to .github ISSUE_TEMPLATE

* add Issue Template Feature Request (#227)

Template for adding features to .github ISSUE_TEMPLATE

* Update Schemas to V3.0RC02 (#207)

Major changes include:
* Update of all schema files to the AAS metamodel version 3.0RC02
* Replacement of most (manually created) examples by new (automatically generated) ones
* Extensive extensions to the explanations in the Readme files
* Introducing YAML files for the AAS Part 2 API definitions

* Update Readmes in schema examples (#242)

The readmes were outdated and referred to the previous examples which
were manually written. We automatically generate the examples now, so we
update the readmes accordingly.

At this point, to make the readmes consistent, we remove the manually
written example for the RDF schema. If we eventually want to restore it,
please do not forget to update the corresponding readme.

Fixes #241.

* Fix abbreviations in labels of RDF schema (#243)

The abbreviations in the labels of RDF schema were not properly
upper-cased due to an omission in aas-core-meta and aas-core-codegen.

We fixed the issue in aas-core-codegen, and propagate the changes in
this patch.

We used [aas-core-codegen 1cff35bd] and [aas-core-meta d180dbb].

[aas-core-codegen 1cff35bd]: https://github.com/aas-core-works/aas-core-codegen/commit/1cff35bd
[aas-core-meta d180dbb]: https://github.com/aas-core-works/aas-core-meta/commit/d180dbb

---------

Co-authored-by: Sebastian Bader <sebastian.r.bader@web.de>
Co-authored-by: Marko Ristin <marko.ristin@gmail.com>
Co-authored-by: Marko Ristin <marko@ristin.ch>